+++ src/main/java/org/loverde/geographiccoordinate/GeographicCoordinate.java
@@ -11,6 +11,7 @@ public interface GeographicCoordinate {
    public int getDegrees();
    public int getMinutes();
    public double getSeconds();
+   public IDirection getDirection();
 
    public double toDouble();
    public double toRadians();