+++ src/main/java/org/loverde/geographiccoordinate/IDirection.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (C) 2016 Kurtis LoVerde
+ * All rights reserved
+ */
+
+
+package org.loverde.geographiccoordinate;
+
+
+public interface IDirection {
+
+   public String getAbbreviation();
+}